Rename collection 'ready' tag to 'published'
The thinking behind the term 'ready' was that records in such a collection are not *necessarily* published - actual record publication might depend on record QC, for example. But this tag is the key criterion for deciding whether or not to publish any record in a collection. If a collection *does not* have this tag, then we know that none of its records will be published. On the contrary, if a collection *does* have this tag, then we know that its records are *expected* to be automatically published - possibly depending on further criteria. (In fact, in the cases of migrated or harvested collections, there are usually no further criteria).