Add scopes and endpoints for reading all packages/resources
This change adds package and resource scopes that are not constrainable by provider - for use by the admin UI, so that SAEON staff can see all packages/resources regardless of their user-provider association (if any). Provider-constrainment of package and resource read/write scopes is intended for end-user interfaces for package and resource upload.

The package and resource admin scopes are also adjusted to be non-constrainable. Still todo: create additional API routes to allow writing of any package/resource by admins.