Update institution keyword schema
Based on discussions with the curation team, the institution title - rather than the acronym/abbreviation - will be the keyword. The exemplary case is "Oceans and Coastal Research, Department of Forestry, Fisheries and the Environment (DFFE)" in which the subdivision (Oceans and Coastal branch of the DFFE) shares an acronym with its parent institution.

We have additionally taken the decision not to define institutions hierarchically. This forestalls a potentially endless nightmare of trying to correctly represent institutional hierarchies, which in many cases (e.g. overseas institutions) would be difficult.